[fix] fixed bug deleting objects from DB, subject to hierarchical structure

+++ src/main/java/com/example/application/service/chapter/ChapterServiceImpl.java
@@ -1,6 +1,8 @@
 package com.example.application.service.chapter;
 
 import com.example.application.dao.chapter.ChapterDAO;
+import com.example.application.dao.section.SectionDAO;
+import com.example.application.dao.textstory.TextStoryDAO;
 import com.example.application.model.Chapter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,10 @@ public class ChapterServiceImpl implements ChapterService {
 
     @Autowired
     private ChapterDAO chapterDAO;
+    @Autowired
+    private SectionDAO sectionDAO;
+    @Autowired
+    private TextStoryDAO textStoryDAO;
 
     @Override
     @Transactional
@@ -35,6 +41,17 @@ public class ChapterServiceImpl implements ChapterService {
     @Override
     @Transactional
     public void deleteChapter(int id) {
-        chapterDAO.deleteChapter(id);
+        Chapter chapterToDelete = getChapter(id);
+        if (chapterToDelete != null) {
+            if(chapterToDelete.getSections() != null && chapterToDelete.getSections().size() > 0) {
+                chapterToDelete.getSections().forEach(section -> {
+                    sectionDAO.deleteSection(section.getId());
+                });
+                chapterDAO.deleteChapter(id);
+                if(chapterToDelete.getText() != null) {
+                    textStoryDAO.deleteTextStory(chapterToDelete.getText().getId());
+                }
+            }
+        }
     }
 }
