+++ src/main/java/com/example/application/service/project/ProjectServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.application.service.project;
 
+import com.example.application.dao.book.BookDAO;
 import com.example.application.dao.project.BookProjectDAO;
 import com.example.application.model.BookProject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +14,8 @@ public class ProjectServiceImpl implements ProjectService {
 
     @Autowired
     private BookProjectDAO bookProjectDAO;
+    @Autowired
+    private BookDAO bookDAO;
 
 
     @Override
@@ -37,6 +40,14 @@ public class ProjectServiceImpl implements ProjectService {
     @Override
     @Transactional
     public void deleteProject(int id) {
-        bookProjectDAO.deleteProject(id);
+        BookProject projectToDelete = getProject(id);
+        if (projectToDelete != null) {
+            if (projectToDelete.getBooks() != null && projectToDelete.getBooks().size() > 0) {
+                projectToDelete.getBooks().forEach(book -> {
+                    bookDAO.deleteBook(book.getId());
+                });
+            }
+            bookProjectDAO.deleteProject(id);
+        }
     }
 }
